+++ src/modules/fileformats/ngf.js
@@ -122,7 +122,7 @@ exports.parse = function(content) {
 
         let tmp = new Board(boardsize, boardsize)       // Created solely for .getHandicapPlacement()
 
-        let points = tmp.getHandicapPlacement(handicap)
+        let points = tmp.getHandicapPlacement(handicap, true)
 
         for (let p of points) {
             let [x, y] = p